requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
amqp==2.4.2               # via kombu
backcall==0.1.0           # via ipython
billiard==3.6.0.0         # via celery
celery[redis]==4.3.0
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
decorator==4.3.2          # via ipython, traitlets
django-anymail[mailgun]==6.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==2.1.0
django==2.2
djangorestframework==3.9.2
idna==2.8                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.3.0
jedi==0.13.3              # via ipython
kombu==4.5.0              # via celery
parso==0.3.4              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.4.0
prompt-toolkit==2.0.9     # via ipython
psycopg2==2.7.7
ptyprocess==0.6.0         # via pexpect
pygments==2.3.1           # via ipython
pytz==2018.9              # via celery, django
redis==3.2.1
requests==2.21.0          # via django-anymail
six==1.12.0               # via django-anymail, pip-tools, prompt-toolkit, traitlets
sqlparse==0.2.4           # via django, django-debug-toolbar
traitlets==4.3.2          # via ipython
urllib3==1.24.2           # via requests
vine==1.3.0               # via amqp, celery
wcwidth==0.1.7            # via prompt-toolkit
